requirements.txt
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
google-auth==1.11.2
google-oauth==1.0.1
grpcio==1.27.2
grpcio-tools==1.27.2
idna==2.9
protobuf==3.11.3
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==19.1.0
requests==2.23.0
rsa==4.0
six==1.14.0
urllib3==1.25.8
